Add support for optional txn parameters.
Makes a number of adjustments to allow the specification of 'optional'
transaction parameters, i.e. those that would otherwise be given a value
by a node if omitted.  These include the gas price, gas limit, and block
information.

These can be specified by running an Azimuth action with runAzimuth'
(note the apostrophe) instead of runAzimuth.